requirements.txt
streamlit>=1.32.0
openai>=1.50.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pymupdf>=1.24.0
pytest>=8.0.0
